#422 fix product create + start adjusting shopify importer
