+++ src/core/integrations/integrations/integrations-show/containers/imports/containers/create-import/containers/general/languages-and-currencies/LanguagesAndCurrenciesStep.vue
@@ -1,19 +1,19 @@
 <script setup lang="ts">
 import { onMounted, ref, defineProps, computed, defineExpose, watch } from "vue";
 import { useI18n } from "vue-i18n";
-import { remoteLanguagesQuery, remoteCurrenciesQuery } from "../../../../../../../../../../shared/api/queries/salesChannels.js";
-import {companyLanguagesQuery} from "../../../../../../../../../../shared/api/queries/languages.js";
-import {currenciesQuery} from "../../../../../../../../../../shared/api/queries/currencies.js";
+import { remoteLanguagesQuery, remoteCurrenciesQuery } from "../../../../../../../../../../../shared/api/queries/salesChannels.js";
+import {companyLanguagesQuery} from "../../../../../../../../../../../shared/api/queries/languages.js";
+import {currenciesQuery} from "../../../../../../../../../../../shared/api/queries/currencies.js";
 
-import {FieldType} from "../../../../../../../../../../shared/utils/constants";
-import apolloClient from "../../../../../../../../../../../apollo-client";
+import {FieldType} from "../../../../../../../../../../../shared/utils/constants";
+import apolloClient from "../../../../../../../../../../../../apollo-client";
 import {
   FieldQuery
-} from "../../../../../../../../../../shared/components/organisms/general-form/containers/form-fields/field-query";
-import { currencyOnTheFlyConfig } from "../../../../../../../../../settings/currencies/configs";
-import { DiscreteLoader } from "../../../../../../../../../../shared/components/atoms/discrete-loader";
-import { RemoteLanguage, RemoteCurrency } from "../../../../configs";
-import { QueryFormField } from "../../../../../../../../../../shared/components/organisms/general-form/formConfig";
+} from "../../../../../../../../../../../shared/components/organisms/general-form/containers/form-fields/field-query";
+import { currencyOnTheFlyConfig } from "../../../../../../../../../../settings/currencies/configs";
+import { DiscreteLoader } from "../../../../../../../../../../../shared/components/atoms/discrete-loader";
+import { RemoteLanguage, RemoteCurrency } from "../../../../../configs";
+import { QueryFormField } from "../../../../../../../../../../../shared/components/organisms/general-form/formConfig";
 
 const { t } = useI18n();
