+++ src/core/integrations/integrations/integrations-show/containers/imports/containers/create-import/containers/magento/default-rule/DefaultRuleStep.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import { ref, computed, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
-import { Toggle } from '../../../../../../../../../../shared/components/atoms/toggle';
-import { Button } from '../../../../../../../../../../shared/components/atoms/button';
-import { Icon } from '../../../../../../../../../../shared/components/atoms/icon';
+import { Toggle } from '../../../../../../../../../../../shared/components/atoms/toggle';
+import { Button } from '../../../../../../../../../../../shared/components/atoms/button';
+import { Icon } from '../../../../../../../../../../../shared/components/atoms/icon';
 
 const props = defineProps<{
   attributeSets: any[];
