+++ src/core/integrations/integrations/integrations-show/containers/imports/containers/create-import/containers/magento/default-rule/index.ts
