+++ src/core/integrations/integrations/integrations-show/containers/imports/containers/create-import/containers/magento/magento-importer/index.ts
@@ -1,0 +1,1 @@
+export { default as MagentoImporter } from "./MagentoImporter.vue";
